extract common utils to separate file
